separar modulos e trocar doubles para float

+++ CG/common/point.h
@@ -1,9 +1,9 @@
 #ifndef point_h
 #define point_h
 
-typedef struct point { double p[3]; } Point;
+typedef struct point { float p[3]; } Point;
 
-Point* mkPoint(double x, double y, double z);
+Point* mkPoint(float x, float y, float z);
 void unmkPoint(Point* p);
 
 #endif
